Init gio hang

+++ src/java/dao/DAO.java
@@ -33,7 +33,7 @@ public class DAO {
                 list.add(new Product(rs.getInt(1),
                         rs.getString(2),
                         rs.getString(3),
-                        rs.getDouble(4),
+                        rs.getInt(4),
                         rs.getString(5),
                         rs.getString(6),
                         rs.getInt(9)));
@@ -70,7 +70,7 @@ public class DAO {
                 return new Product(rs.getInt(1),
                         rs.getString(2),
                         rs.getString(3),
-                        rs.getDouble(4),
+                        rs.getInt(4),
                         rs.getString(5),
                         rs.getString(6),rs.getInt(9));
             }
@@ -91,7 +91,7 @@ public class DAO {
                     list.add(new Product(rs.getInt(1),
                             rs.getString(2),
                             rs.getString(3),
-                            rs.getDouble(4),
+                            rs.getInt(4),
                             rs.getString(5),
                             rs.getString(6),rs.getInt(9)));
                 }
@@ -111,7 +111,7 @@ public class DAO {
                     return new Product(rs.getInt(1),
                             rs.getString(2),
                             rs.getString(3),
-                            rs.getDouble(4),
+                            rs.getInt(4),
                             rs.getString(5),
                             rs.getString(6),rs.getInt(9));
                 }
@@ -131,7 +131,7 @@ public class DAO {
                     list.add(new Product(rs.getInt(1),
                             rs.getString(2),
                             rs.getString(3),
-                            rs.getDouble(4),
+                            rs.getInt(4),
                             rs.getString(5),
                             rs.getString(6),rs.getInt(9)));
                 }
@@ -152,7 +152,7 @@ public class DAO {
                     list.add(new Product(rs.getInt(1),
                             rs.getString(2),
                             rs.getString(3),
-                            rs.getDouble(4),
+                            rs.getInt(4),
                             rs.getString(5),
                             rs.getString(6),rs.getInt(9)));
                 }
@@ -309,7 +309,7 @@ public class DAO {
                     list.add(new Product(rs.getInt(1),
                             rs.getString(2),
                             rs.getString(3),
-                            rs.getDouble(4),
+                            rs.getInt(4),
                             rs.getString(5),
                             rs.getString(6),rs.getInt(9)));
                 }
@@ -330,7 +330,7 @@ public class DAO {
                     list.add(new Product(rs.getInt(1),
                             rs.getString(2),
                             rs.getString(3),
-                            rs.getDouble(4),
+                            rs.getInt(4),
                             rs.getString(5),
                             rs.getString(6),rs.getInt(9)));
                 }
